Reconfigure tests after updating to jsdom 11.12.0
jsdom v11.12.0 added support for localStorage apis. Due to that, it is
now a requirement to set an origin url on the jsdom instance if you at
all access the localStorage variable.

Both our custom mocha setup and jest will loop over any of the
properties on window and make them available on global, thus invoking
the getter and triggering the problem.

Setting the origin url to http://localhost solved the problem.

 - https://github.com/facebook/jest/issues/6766
 - https://github.com/jsdom/jsdom#simple-options